fixed readmes and aws policies
