+++ aws/data_aws_assume_role_policy.go
@@ -1,0 +1,91 @@
+package aws
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
+
+type awsIamPolicy struct {
+	Version    string                   `json:"Version,omitempty"`
+	ID         string                   `json:"Id,omitempty"`
+	Statements []*awsIamPolicyStatement `json:"Statement"`
+}
+
+type awsIamPolicyStatement struct {
+	Sid          string                       `json:"Sid,omitempty"`
+	Effect       string                       `json:"Effect,omitempty"`
+	Actions      interface{}                  `json:"Action,omitempty"`
+	NotActions   interface{}                  `json:"NotAction,omitempty"`
+	Resources    interface{}                  `json:"Resource,omitempty"`
+	NotResources interface{}                  `json:"NotResource,omitempty"`
+	Principal    map[string]string            `json:"Principal,omitempty"`
+	Condition    map[string]map[string]string `json:"Condition,omitempty"`
+}
+
+
+// DataAwsAssumeRolePolicy ...
+func DataAwsAssumeRolePolicy() *schema.Resource {
+	return &schema.Resource{
+		ReadContext: func(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+			externalID := d.Get("external_id").(string)
+			policy := awsIamPolicy{
+				Version: "2012-10-17",
+				Statements: []*awsIamPolicyStatement{
+					{
+						Effect:  "Allow",
+						Actions: "sts:AssumeRole",
+						Condition: map[string]map[string]string{
+							"StringEquals": {
+								"sts:ExternalId": externalID,
+							},
+						},
+						Principal: map[string]string{
+							"AWS": fmt.Sprintf("arn:aws:iam::%s:root", d.Get("databricks_account_id").(string)),
+						},
+					},
+				},
+			}
+			if v, ok := d.GetOk("for_log_delivery"); ok {
+				if v.(bool) {
+					// this is production UsageDelivery IAM role, that is considered a constant
+					logDeliveryARN := "arn:aws:iam::414351767826:role/SaasUsageDeliveryRole-prod-IAMRole-3PLHICCRR1TK"
+					policy.Statements[0].Principal["AWS"] = logDeliveryARN
+				}
+			}
+			policyJSON, err := json.MarshalIndent(policy, "", "  ")
+			if err != nil {
+				return diag.FromErr(err)
+			}
+			d.SetId(externalID)
+			// nolint
+			d.Set("json", string(policyJSON))
+			return nil
+		},
+		Schema: map[string]*schema.Schema{
+			"databricks_account_id": {
+				Type:     schema.TypeString,
+				Default:  "414351767826",
+				Optional: true,
+			},
+			"for_log_delivery": {
+				Type:        schema.TypeBool,
+				Description: "Grant AssumeRole to Databricks SaasUsageDeliveryRole instead of root account",
+				Optional:    true,
+				Default:     false,
+			},
+			"external_id": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"json": {
+				Type:     schema.TypeString,
+				Computed: true,
+				ForceNew: true,
+			},
+		},
+	}
+}
